Initial Commit
Read Me

The code is written in 2.7

In order to run the program just copy paste the following when your
prompt is in the same directory as the python file:

cd downloads
python simple_lesk_imp.py

It will ask for a sentence once provided with a sentence it will for
the word to disambiguate. Once provided it will give the most sensible
definition as the out put.

Documentation:

The examples of the words are :

Bank , plant, ran, mug

Example sentences where it will be a success:

I go to the river bank to find peace.

He was a plant into the gang.

Example sentences where it will be a failure:

I go to the bank on the bank of river nile to deposit money.